move extract_spectrum_from_image to extractor.py
